requirements.txt
Faker==37.4.0
iniconfig==2.1.0
kafka-python==2.2.15
packaging==25.0
pluggy==1.6.0
prometheus_client==0.22.1
Pygments==2.19.2
pytest==8.4.1
tzdata==2025.2
zstandard==0.23.0
